requirements.txt
pyyaml
openai
# lightrag-hku  # Example: pip install git+https://github.com/HKUDS/LightRAG.git
# rag-anything  # Example: pip install git+https://github.com/HKUDS/RAG-Anything.git

